refactor: 패키지 구조를 수정하라
