+++ src/main/java/racingcar/view/output/OutputView.java
@@ -1,0 +1,20 @@
+package racingcar.view.output;
+
+public class OutputView {
+    public static final String ERROR_HEADER_MESSAGE = "[ERROR]";
+
+    private OutputView() {}
+
+    public static void printErrorMessage(String message) {
+        System.out.println(ERROR_HEADER_MESSAGE + " " + message);
+    }
+
+    public static void carsRaceStatus(Cars cars) {
+        for (Car car : cars.getCars()) {
+            raceStatus(car);
+        }
+    }
+    private static void raceStatus(Car car) {
+        System.out.println(car.getName() + ": " + car.getDistanceByDash());
+    }
+}
